# packages/api-contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/verification/session": {
      "post": {
        "operationId": "IdentityController_startSession",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartVerificationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Identity"]
      }
    },
    "/api/v1/verification/session/{id}": {
      "get": {
        "operationId": "IdentityController_session",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Identity"]
      }
    },
...
    "/api/v1/verification/recovery-requests/{id}/approve": {
      "post": {
        "operationId": "IdentityController_approveRecovery",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveRecoveryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Identity"]
      }
    },
    "/api/v1/verification/recovery-requests/{id}/reject": {
      "post": {
        "operationId": "IdentityController_rejectRecovery",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectRecoveryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Identity"]
      }
    },
...
    "/api/v1/bookings/{id}/disputes": {
      "post": {
        "operationId": "DisputesController_open",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenDisputeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Disputes"]
      },
      "get": {
        "operationId": "DisputesController_list",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Disputes"]
      }
    },
    "/api/v1/disputes/{id}/resolve": {
      "post": {
        "operationId": "DisputesController_resolve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveDisputeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Disputes"]
      }
    },
...
    "/api/v1/booking-requests/{id}/match": {
      "post": {
        "operationId": "MatchingController_match",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Matching"]
      },
      "get": {
        "operationId": "MatchingController_result",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Matching"]
      }
    },
    "/api/v1/matching/runs": {
      "post": {
        "operationId": "MatchingController_batch",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchBatchDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Matching"]
      }
    },
    "/api/v1/matching/runs/{requestId}": {
      "get": {
        "operationId": "MatchingController_run",
        "parameters": [
          {
            "name": "requestId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Matching"]
      }
    },
...
  },
  "info": {
    "title": "Emek Core API",
    "description": "İki taraflı dijital hizmet pazaryeri çekirdek API. Hata kodları: docs/api/error-codes.md",
    "version": "v1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Firebase ID token (ADR-0016)"
      }
    },
    "schemas": {
...
      "StartVerificationDto": {
        "type": "object",
        "properties": {}
      },
      "ApproveRecoveryDto": {
        "type": "object",
        "properties": {}
      },
      "RejectRecoveryDto": {
        "type": "object",
        "properties": {}
      },
...
      "OpenDisputeDto": {
        "type": "object",
        "properties": {}
      },
      "ResolveDisputeDto": {
        "type": "object",
        "properties": {}
      },
...
      "MatchBatchDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}